ADDED universal_sdk automated validation for removal during syn command
