+++ cli/src/core/ast.rs
@@ -4,3 +4,4 @@ pub(crate) mod injections;
 pub(crate) mod parse_ast_program;
 pub(crate) mod replacements;
 pub(crate) mod transformations;
+pub(crate) mod validation;
